pyproject.toml
[tool.poetry]
name = "reusable-services"
version = "0.1.0"
description = "A Reusable Services writting in Various Languages."
authors = ["Hiroaki Yamamoto <hiroaki@hysoftware.net>"]
license = "MIT"
repository = "https://github.com/hiroaki-yamamoto/reusable-services.git"
keywords = ["microservice", "framework"]
classifiers = [
  "Development Status :: 1 - Planning",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Go",
]
packages = [
  { include = "render/**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.7, ^3.8"
jinja2 = "^2.11.1"
grpcio-tools = "^1.27.2"
msgpack = "^1.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
flake8-docstrings = "^1.5.0"
flake8-commas = "^2.0.0"
tox = "^3.14.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
